+++ packages/oauth/oauth-client/src/oauth-client.ts
@@ -43,7 +43,7 @@ import type { RuntimeImplementation } from './runtime-implementation.js'
 import { Runtime } from './runtime.js'
 import {
   SessionGetter,
-  type SessionHooks,
+  type SessionGetterOptions,
   type SessionStore,
   isExpectedSessionError,
 } from './session-getter.js'
@@ -70,7 +70,7 @@ export type {
   OAuthResponseMode,
   ProtectedResourceMetadataCache,
   RuntimeImplementation,
-  SessionHooks,
+  SessionGetterOptions,
   SessionStore,
   StateStore,
 }
@@ -109,7 +109,7 @@ export type OAuthClientOptions = {
   runtimeImplementation: RuntimeImplementation
   fetch?: Fetch
 } & CreateIdentityResolverOptions &
-  SessionHooks
+  SessionGetterOptions
 
 export type OAuthClientFetchMetadataOptions = {
   clientId: OAuthClientIdDiscoverable