+++ packages/oauth/oauth-provider/src/request/request-manager.ts
@@ -254,11 +254,13 @@ export class RequestManager {
       )
     }
 
-    // atproto extension: if the client is not trusted, and not authenticated,
+    // atproto extension: if the client is not trusted, and not confidential,
     // force users to consent to authorization requests. We do this to avoid
     // unauthenticated clients from being able to silently re-authenticate
     // users.
-    if (!client.isTrusted && !client.isFirstParty && !client.isConfidential) {
+    // @NOTE First party clients are trusted by default, but they can
+    // also be explicitly marked as un-trusted.
+    if (!client.isConfidential && !client.isTrusted) {
       if (parameters.prompt === 'none') {
         throw new ConsentRequiredError(
           parameters,