+++ packages/dev-env/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @atproto/dev-env
 
+## 0.5.39
+
+### Patch Changes
+
+- Updated dependencies [[`102503c`](https://github.com/bluesky-social/atproto/commit/102503c0781e3605ac725adf5c94893870f7d2f7), [`5c5df26`](https://github.com/bluesky-social/atproto/commit/5c5df26bdeed22a06aa08e0463b76ad19be59ed8)]:
+  - @atproto/bsky@0.0.267
+  - @atproto/api@0.20.34
+  - @atproto/ozone@0.2.21
+
+## 0.5.38
+
+### Patch Changes
+
+- Updated dependencies [[`5d6dae2`](https://github.com/bluesky-social/atproto/commit/5d6dae261032033eb4c818944208b9342c1d49d4)]:
+  - @atproto/bsky@0.0.266
+  - @atproto/api@0.20.33
+  - @atproto/ozone@0.2.20
+
 ## 0.5.37
 
 ### Patch Changes